requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2
alembic
pydantic
passlib
python-jose
redis